refactored project architecture.
